Address PR #201 review findings
- Restore the preset round-trip test's deletion assertion that was
  accidentally absorbed by the new malformed-preset test
- /api/output: keep a deprecated output alias on cursor-less requests
  so external consumers of the pre-cursor contract keep working (the
  response key was renamed to lines in the cursor protocol)
- Deep-copy runtime health cache entries so callers cannot mutate
  nested lists inside cached results
- stop_process: keep tracking a process that survives kill() instead of
  clearing state and allowing a concurrent second launch
- launch_process: kill and untrack a partially launched process if
  state wiring fails after Popen succeeds
- Content-Length: require ASCII digits so the lexicographic size check
  stays correct for Unicode decimal digits
- web_search: fix timeout annotations (int -> float) and document that
  bypassing system proxies is intentional for SSRF IP pinning

+++ backend/services/web_search.py
@@ -132,7 +132,7 @@ class ManualRedirectHandler(urllib.request.HTTPRedirectHandler):
 NoRedirect = ManualRedirectHandler
 
 
-def _connect_validated(addresses: list[Any], timeout: int, source_address: Any = None) -> Any:
+def _connect_validated(addresses: list[Any], timeout: float, source_address: Any = None) -> Any:
     last_error = None
     for family, socktype, proto, _, sockaddr in addresses:
         sock = socket.socket(family, socktype, proto)
@@ -151,7 +151,7 @@ def _connect_validated(addresses: list[Any], timeout: int, source_address: Any =
 
 
 class _PinnedHTTPConnection(http.client.HTTPConnection):
-    def __init__(self, host: str, port: int, addresses: list[Any], timeout: int) -> None:
+    def __init__(self, host: str, port: int, addresses: list[Any], timeout: float) -> None:
         self._validated_addresses = addresses
         super().__init__(host, port=port, timeout=timeout)
 
@@ -169,7 +169,7 @@ class _PinnedHTTPSConnection(http.client.HTTPSConnection):
         host: str,
         port: int,
         addresses: list[Any],
-        timeout: int,
+        timeout: float,
         ssl_context: Optional[Any],
     ) -> None:
         self._validated_addresses = addresses
@@ -187,9 +187,11 @@ class _PinnedHTTPSConnection(http.client.HTTPSConnection):
 def _open_validated_url(
     parsed: urllib.parse.ParseResult,
     addresses: list[Any],
-    timeout: int,
+    timeout: float,
     ssl_context: Optional[Any],
 ) -> tuple[int, str, Any, bytes]:
+    # Connects directly to the pre-validated addresses (no system proxy
+    # support): routing through a proxy would bypass the SSRF IP pinning.
     port = parsed.port or (443 if parsed.scheme == "https" else 80)
     connection_class = _PinnedHTTPSConnection if parsed.scheme == "https" else _PinnedHTTPConnection
     if parsed.scheme == "https":
